fix: setup watcher directly from new InteractiveFeaturesLayer component
Watchers which are added asynchronously in script setup are not cleaned automatically when component is unmounted, which leads to errors when we change feature store selection when the map has been unmounted.

Setting watchers synchronously directly in script setup of InteractiveFeaturesLayer ensures watchers are added and removed at the same time as features are added and removed from the map.

Signed-off-by: Maud Royer <hello@maudroyer.fr>